fix(types): audit and fix remaining long → int32_t issues (spec-002)
Fix critical struct layout mismatch between source/build.h (game code)
and SRC/BUILD.H (engine code). On 64-bit Linux, long is 8 bytes while
int32_t is 4 bytes, causing a 56 vs 40 byte sectortype struct size
mismatch — every sector[i] access with i>0 would hit wrong memory.

source/build.h:
- Change sectortype, walltype, spritetype fields from long to int32_t
- Add #pragma pack(push,1) / pack(pop) for binary format compatibility
- Add _Static_assert checks matching SRC/BUILD.H
- Add #include <stdint.h>

Engine function signatures (SRC/BUILD.H, SRC/ENGINE.C):
- clipmove, pushmove: long *x,*y,*z → int32_t *x,*y,*z
- loadboard, saveboard: long *dapos{x,y,z} → int32_t *dapos{x,y,z}
- rotatepoint: long *x2,*y2 → int32_t *x2,*y2

source/DUKE3D.H, source/GLOBAL.C:
- player_struct: posx,posy,posz,oposx,oposy,oposz → int32_t
- Global vars: myx,myy,myz,omyx,omyy,omyz,myxvel,myyvel,myzvel → int32_t
- Arrays: myxbak,myybak,myzbak → int32_t

source/ACTORS.C:
- Replace dangerous (long *)&sc->ceilingz pattern with typed int32_t *zptr
- Fix local vars: daz,oldx,oldy,gutz,floorz,dax,day,tx,ty → int32_t
- Use temp int32_t vars for rotatepoint output

source/SECTOR.C, source/PLAYER.C:
- Fix local vars storing struct field values: dax,day → int32_t
- Fix goalx,goaly,goalz arrays → int32_t
- Fix computergetinput position vars → int32_t

+++ SRC/ENGINE.C
@@ -2275,7 +2275,7 @@ transmaskwallscan(long x1, long x2)
 	faketimerhandler();
 }
 
-loadboard(char *filename, long *daposx, long *daposy, long *daposz,
+loadboard(char *filename, int32_t *daposx, int32_t *daposy, int32_t *daposz,
 			 short *daang, short *dacursectnum)
 {
 	short fil, i, numsprites;
@@ -2319,7 +2319,7 @@ loadboard(char *filename, long *daposx, long *daposy, long *daposz,
 	return(0);
 }
 
-saveboard(char *filename, long *daposx, long *daposy, long *daposz,
+saveboard(char *filename, int32_t *daposx, int32_t *daposy, int32_t *daposz,
 			 short *daang, short *dacursectnum)
 {
 	short fil, i, j, numsprites;
@@ -5395,7 +5395,7 @@ lastwall(short point)
 }                                                        \
 
 long clipmoveboxtracenum = 3;
-clipmove (long *x, long *y, long *z, short *sectnum,
+clipmove (int32_t *x, int32_t *y, int32_t *z, short *sectnum,
 			 long xvect, long yvect,
 			 long walldist, long ceildist, long flordist, unsigned long cliptype)
 {
@@ -5764,7 +5764,7 @@ raytrace (long x3, long y3, long *x4, long *y4)
 	return(hitwall);
 }
 
-pushmove (long *x, long *y, long *z, short *sectnum,
+pushmove (int32_t *x, int32_t *y, int32_t *z, short *sectnum,
 			 long walldist, long ceildist, long flordist, unsigned long cliptype)
 {
 	sectortype *sec, *sec2;
@@ -5932,7 +5932,7 @@ updatesector(long x, long y, short *sectnum)
 	*sectnum = -1;
 }
 
-rotatepoint(long xpivot, long ypivot, long x, long y, short daang, long *x2, long *y2)
+rotatepoint(long xpivot, long ypivot, long x, long y, short daang, int32_t *x2, int32_t *y2)
 {
 	long dacos, dasin;
